Tutorial for Gridded Data Structures
Tutorial with the basic steps required for working with ndarrays in Python for Engineers and Geoscientists.  This includes methods to load and save, search, slice, modify etc.  With these steps one should be able to build Python workflows for statistics, geostatistics, and machine learning-based data analytics with the typical datasets we work with.